pyproject.toml
[tool.poetry]
name = "VectorScan"
version = "0.1.0"
description = "A vector database indexing system for semantic search across different frameworks."
authors = [
   "Oleg Maslo"
]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
click = "*"
faiss-cpu = "*"
sentence-transformers = "*"
pytest = "*"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = ["test_*.py"]
pythonpath = "."